Modal date field now calendar
Заміна поля date зі строки на календарний ввод
Розбиття файлу модалки на складові для зручності читання
